Add `WithHeaderSeparatorRow`
This function will print a separator line between the header and the
data rows. It will use the same formatter as the header. The width of
each separator cell will be equal to the width of the header cell in the
same column. It supports runes, meaning dependning on the rune width, it
might not be perfect.